@@ -48,15 +48,15 @@
   ],
   "dependencies": {
     "@ahmed757/ngx-treant-js": "^1.0.4",
-    "@angular/animations": "~13.4.0",
-    "@angular/common": "~13.4.0",
-    "@angular/compiler": "~13.4.0",
-    "@angular/core": "~13.4.0",
-    "@angular/forms": "^13.4.0",
-    "@angular/localize": "~13.4.0",
-    "@angular/platform-browser": "~13.4.0",
-    "@angular/platform-browser-dynamic": "~13.4.0",
-    "@angular/router": "~13.4.0",
+    "@angular/animations": "^14.3.0",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/localize": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "bootstrap": "4.5.0",
     "jquery": "3.5.1",
     "jquery-easing": "0.0.1",
@@ -69,11 +69,11 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.11",
-    "@angular-devkit/build-ng-packagr": "~0.1002.4",
-    "@angular/cli": "~13.3.11",
-    "@angular/compiler-cli": "~13.4.0",
-    "@angular/language-service": "~13.4.0",
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular-devkit/build-ng-packagr": "^0.1002.4",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/language-service": "^14.3.0",
     "@commitlint/cli": "^9.0.1",
     "@commitlint/config-conventional": "^9.0.1",
     "@types/jasmine": "~3.5.0",
@@ -92,7 +92,7 @@
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
     "lint-staged": "^10.2.11",
-    "ng-packagr": "^13.3.1",
+    "ng-packagr": "^14.2.2",
     "prettier": "2.0.5",
     "protractor": "~7.0.0",
     "semantic-release": "^17.1.1",
